+++ NERBABO.Backend/NERBABO.ApiService/Core/Global/Dtos/RetrieveGeneralInfoDto.cs
@@ -11,7 +11,7 @@ public class RetrieveGeneralInfoDto
     public string Nipc { get; set; } = string.Empty;
     public string LogoFinancing { get; set; } = string.Empty;
     public int IvaId { get; set; }
-    public int IvaPercent { get; set; }
+    public float IvaPercent { get; set; }
     public string IvaRegime { get; set; } = string.Empty;
 
 }
