+++ NERBABO.Backend/NERBABO.ApiService/Core/Global/Dtos/RetrieveTaxDto.cs
@@ -4,7 +4,7 @@ public class RetrieveTaxDto
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public int ValuePercent { get; set; }
+    public float ValuePercent { get; set; }
     public float ValueDecimal => (float)ValuePercent / 100;
     public bool IsActive { get; set; }
     public string Type { get; set; } = string.Empty;
