+++ NERBABO.Backend/NERBABO.ApiService/Core/Global/Dtos/UpdateTaxDto.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace NERBABO.ApiService.Core.Global.Dtos;
 
-public class UpdateTaxDto: EntityDto<int>
+public class UpdateTaxDto : EntityDto<int>
 {
     [Required(ErrorMessage = "Regime é um campo obrigatório")]
     [StringLength(50, MinimumLength = 3, ErrorMessage = "Regime deve conter pelo menos {2} caracteres e um máximo de {1} caracteres")]
@@ -11,7 +11,7 @@ public class UpdateTaxDto: EntityDto<int>
 
     [Required(ErrorMessage = "Valor percentual é um campo obrigatório")]
     [Range(0, 100, ErrorMessage = "Valor percentual deve estar entre {1} e {2}")]
-    public int ValuePercent { get; set; }
+    public float ValuePercent { get; set; }
     public bool IsActive { get; set; }
 
     [Required(ErrorMessage = "Tipo é um campo obrigatório")]
